Add Processed Data
+ backlink to processor repo (`in/AlSTJ-Gen2`)
+ processed data

This turns this repo into a cyclical store of data from this experiment in any form: raw or processed. Whereas before I envisioned it holding only experimental related content, such as photos, lab notes, the data, and metadata, this now introduces data created *after* the experiment.

The reason is simply because I needed a place to store large data on a filesystem that had the bandwidth for it, which is the case on both the LC and CRONOS. So, I think this precedent can be set.

In practice, this means adding as a submodule any repos that import data into this repo. I think `in/` is fine for this, but there is one caveat for users. Definitely DO NOT initialize the copy of this repo that is referenced in the importer submodule (there's basically no need to init the importer submodules at all unless you want to see the code that made the data that's now in this repo).

Given this, I might consider a new root folder called `backlinks/` that are specifically repos that provide files in this repo, where the name is taken from bidirectional link nomenclature. All in all, having bidirectional submodules helps with provenance tracking.

Delete partial files

Confirmed that these are aggregated into files during processing, so these are unnecessary.
Annex dropping them too.