Rename to ISO Date
While the symlinks in in/ are NOT the correct date, that will have to be for another commit.
Already this commit breaks my code which is looking for the wrong date format, so I gotta patch that.
Still pining for the day that my data is actually IN submodules and not actually symlinks..