Move to 6ALY
Needs to be deadened by moving all known copies to the depot at data_fast on CRONOS.
Then, use 6ALY to move the file out of the depot. Once this is done, the key becomes inaccessibly by 5DAT.
Finally, the key may be deadened, which limits any interaction using --all, otherwise it's not that big of a deal.